Use tostring instead of implicit conversion
Apparently when you call a function that returns a uint32_t it turns the result
into an 'anumber', but when you create a uint32_t via ffi.new the result is a
'bcdata'. anumbers convert automatically to string, but bcdata does not. So
while the code was fine, for clarity this code is changed. (Note this also
counts for commit e95f94b)